// apps/webapp/spa/src/lib/api/generated/models/ContentPillar.ts
export interface ContentPillar {
    /**
     * 
     * @type {string}
     * @memberof ContentPillar
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ContentPillar
     */
    name?: string;
    /**
     * 
     * @type {ContentPillarBusinessGoal}
     * @memberof ContentPillar
     */
    businessGoal?: ContentPillarBusinessGoal;
    /**
     * 
     * @type {ContentPillarType}
     * @memberof ContentPillar
     */
    type?: ContentPillarType;
    /**
     * 
     * @type {ContentPillarFunnelStage}
     * @memberof ContentPillar
     */
    funnelStage?: ContentPillarFunnelStage;
    /**
     * 
     * @type {Array<ContentTypeName>}
     * @memberof ContentPillar
     */
    contentTypes?: Array<ContentTypeName>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ContentPillar
     */
    audienceIds?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ContentPillar
     */
    hooks?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ContentPillar
     */
    callsToAction?: Array<string>;
}